Catch shared controller logic in trait
The shared methods from the EntityCreateController and the EntityEditController are now brought together in the EntityControllerTrait.
