+++ src/Surfnet/ServiceProviderDashboard/Infrastructure/DashboardBundle/Controller/EntityControllerTrait.php
@@ -1,0 +1,154 @@
+<?php
+
+/**
+ * Copyright 2017 SURFnet B.V.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+namespace Surfnet\ServiceProviderDashboard\Infrastructure\DashboardBundle\Controller;
+
+use League\Tactician\CommandBus;
+use Surfnet\ServiceProviderDashboard\Application\Command\Entity\DeleteEntityCommand;
+use Surfnet\ServiceProviderDashboard\Application\Command\Entity\LoadMetadataCommand;
+use Surfnet\ServiceProviderDashboard\Application\Command\Entity\PublishEntityProductionCommand;
+use Surfnet\ServiceProviderDashboard\Application\Command\Entity\PublishEntityTestCommand;
+use Surfnet\ServiceProviderDashboard\Application\Command\Entity\SaveEntityCommand;
+use Surfnet\ServiceProviderDashboard\Application\Exception\InvalidArgumentException;
+use Surfnet\ServiceProviderDashboard\Application\Service\EntityService;
+use Surfnet\ServiceProviderDashboard\Application\Service\ServiceService;
+use Surfnet\ServiceProviderDashboard\Domain\Entity\Entity;
+use Surfnet\ServiceProviderDashboard\Infrastructure\DashboardBundle\Factory\MailMessageFactory;
+use Surfnet\ServiceProviderDashboard\Infrastructure\DashboardBundle\Form\Entity\EntityType;
+use Surfnet\ServiceProviderDashboard\Infrastructure\DashboardBundle\Service\AuthorizationService;
+use Surfnet\ServiceProviderDashboard\Legacy\Metadata\Exception\MetadataFetchException;
+use Surfnet\ServiceProviderDashboard\Legacy\Metadata\Exception\ParserException;
+use Symfony\Component\Form\FormInterface;
+use Symfony\Component\HttpFoundation\RedirectResponse;
+use Symfony\Component\HttpFoundation\Request;
+use Symfony\Component\HttpFoundation\Session\Flash\FlashBagInterface;
+
+/**
+ * The EntityControllerTrait contains the shared logic of the EntityCreateController and the EntityEditController.
+ */
+trait EntityControllerTrait
+{
+    /**
+     * @var CommandBus
+     */
+    private $commandBus;
+
+    /**
+     * @var EntityService
+     */
+    private $entityService;
+
+    /**
+     * @var ServiceService
+     */
+    private $serviceService;
+
+    /**
+     * @var AuthorizationService
+     */
+    private $authorizationService;
+
+    /**
+     * @var MailMessageFactory
+     */
+    private $mailMessageFactory;
+
+    /**
+     * @param CommandBus $commandBus
+     * @param EntityService $entityService
+     * @param ServiceService $serviceService
+     * @param AuthorizationService $authorizationService
+     * @param MailMessageFactory $mailMessageFactory
+     */
+    public function __construct(
+        CommandBus $commandBus,
+        EntityService $entityService,
+        ServiceService $serviceService,
+        AuthorizationService $authorizationService,
+        MailMessageFactory $mailMessageFactory
+    ) {
+        $this->commandBus = $commandBus;
+        $this->entityService = $entityService;
+        $this->serviceService = $serviceService;
+        $this->authorizationService = $authorizationService;
+        $this->mailMessageFactory = $mailMessageFactory;
+    }
+
+    /**
+     * @param Request $request
+     * @param SaveEntityCommand $command
+     *
+     * @return FormInterface
+     */
+    private function handleImport(Request $request, SaveEntityCommand $command)
+    {
+        // Handle an import action based on the posted xml or import url.
+        $metadataCommand = new LoadMetadataCommand($command, $request->get('dashboard_bundle_entity_type'));
+        try {
+            $this->commandBus->handle($metadataCommand);
+        } catch (MetadataFetchException $e) {
+            $this->addFlash('error', 'entity.edit.metadata.fetch.exception');
+        } catch (ParserException $e) {
+            $this->addFlash('error', 'entity.edit.metadata.parse.exception');
+        } catch (InvalidArgumentException $e) {
+            $this->addFlash('error', 'entity.edit.metadata.invalid.exception');
+        }
+        return $this->createForm(EntityType::class, $command);
+    }
+
+    /**
+     * @param Entity $entity
+     * @param FlashBagInterface $flashBag
+     * @return RedirectResponse|FormInterface
+     */
+    private function publishEntity(Entity $entity, FlashBagInterface $flashBag)
+    {
+        switch ($entity->getEnvironment()) {
+            case Entity::ENVIRONMENT_TEST:
+                $publishEntityCommand = new PublishEntityTestCommand($entity->getId());
+                $this->commandBus->handle($publishEntityCommand);
+
+                if (!$flashBag->has('error')) {
+                    $this->get('session')->set('published.entity.clone', clone $entity);
+
+                    // Test entities are removed after they've been published to Manage
+                    $deleteCommand = new DeleteEntityCommand($entity->getId());
+                    $this->commandBus->handle($deleteCommand);
+
+                    return $this->redirectToRoute('service_published_test');
+                }
+                break;
+
+            case Entity::ENVIRONMENT_PRODUCTION:
+                $message = $this->mailMessageFactory->buildPublishToProductionMessage($entity);
+                $publishEntityCommand = new PublishEntityProductionCommand($entity->getId(), $message);
+                $this->commandBus->handle($publishEntityCommand);
+                return $this->redirectToRoute('service_published_production', ['id' => $entity->getId()]);
+                break;
+        }
+    }
+
+    private function isImportButtonClicked(Request $request)
+    {
+        $data = $request->get('dashboard_bundle_entity_type', false);
+        if ($data && isset($data['metadata']) && isset($data['metadata']['importButton'])) {
+            return true;
+        }
+        return false;
+    }
+}
